requirements.txt
aiml==0.9.1
AIML-Bot==0.0.3
asn1crypto==0.24.0
beautifulsoup4==4.6.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Flask==1.0.2
Flask-Cors==3.0.4
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
pandas==0.23.4
pkg-resources==0.0.0
pycparser==2.18
PyMySQL==0.9.2
python-aiml==0.9.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
stripe==1.82.1
urllib3==1.23
Werkzeug==0.14.1
